BFBayesFactor-class {BayesFactor}	R Documentation

General S4 class for representing multiple Bayes factor model comparisons, all against the same model

Description

The BFBayesFactor class is a general S4 class for representing models model comparison via Bayes factor.

Usage


## S4 method for signature 'numeric,BFBayesFactor'
e1 / e2

## S4 method for signature 'BFBayesFactor,BFBayesFactor'
e1 / e2

## S4 method for signature 'BFBayesFactor,index,missing,missing'
x[i, j, ..., drop = TRUE]

## S4 method for signature 'BFBayesFactor'
t(x)

## S4 method for signature 'BFBayesFactor'
which.max(x)

## S4 method for signature 'BFBayesFactor'
which.min(x)

## S4 method for signature 'BFBayesFactor'
is.na(x)

## S4 method for signature 'BFBayesFactor,BFodds'
e1 * e2

## S4 method for signature 'BFBayesFactorTop,index,missing,missing'
x[i, j, ...,
  drop = TRUE]

Arguments

e1
Numerator of the ratio

e2
Denominator of the ratio

x
BFBayesFactor object

i
indices indicating elements to extract

j
unused for BFBayesFactor objects

...
further arguments passed to related methods

drop
unused

Details

BFBayesFactor objects can be inverted by taking the reciprocal and can be divided by one another, provided both objects have the same denominator. In addition, the t (transpose) method can be used to invert Bayes factor objects.

The BFBayesFactor class has the following slots defined:

numerator

a list of models all inheriting BFmodel, each providing a single denominator

denominator

a single BFmodel object serving as the denominator for all model comparisons

bayesFactor

a data frame containing information about the comparison between each numerator and the denominator

data

a data frame containing the data used for the comparison

version

character string giving the version and revision number of the package that the model was created in

Examples


## Compute some Bayes factors to demonstrate division and indexing
data(puzzles)
bfs <- anovaBF(RT ~ shape*color + ID, data = puzzles, whichRandom = "ID", progress=FALSE)

## First and second models can be separated; they remain BFBayesFactor objects
b1 = bfs[1]
b2 = bfs[2]
b1

## We can invert them, or divide them to obtain new model comparisons
1/b1
b1 / b2

## Use transpose to create a BFBayesFactorList
t(bfs)

[Package BayesFactor version 0.9.12-4.2 Index]