BFInfo {BayesFactor}	R Documentation

Prints the version information for the BayesFactor package

Description

Prints the version, revision, and date information for the BayesFactor package

Usage


BFInfo(print = TRUE)

Arguments

print
if TRUE, print version information to the console

Details

This function prints the version and revision information for the BayesFactor package.

Value

BFInfo returns a character string containing the version and revision number of the package..

Author(s)

Richard D. Morey (richarddmorey@gmail.com)

[Package BayesFactor version 0.9.12-4.2 Index]