BFodds-class {BayesFactor}	R Documentation

General S4 class for representing multiple odds model comparisons, all against the same model

Description

The BFodds class is a general S4 class for representing models model comparison via prior or posterior odds.

Usage


## S4 method for signature 'numeric,BFodds'
e1 / e2

## S4 method for signature 'BFodds,BFodds'
e1 / e2

## S4 method for signature 'BFodds,BFBayesFactor'
e1 * e2

## S4 method for signature 'BFodds,index,missing,missing'
x[i, j, ..., drop = TRUE]

Arguments

e1
Numerator of the ratio

e2
Denominator of the ratio

x
BFodds object

i
indices indicating elements to extract

j
unused for BFodds objects

...
further arguments passed to related methods

drop
unused

Details

BFodds objects can be inverted by taking the reciprocal and can be divided by one another, provided both objects have the same denominator. In addition, the t (transpose) method can be used to invert odds objects.

The BFodds class has the following slots defined:

numerator

a list of models all inheriting BFmodel, each providing a single numerator

denominator

a single BFmodel object serving as the denominator for all model comparisons

logodds

a data frame containing information about the (log) prior odds between each numerator and the denominator

bayesFactor

a BFBayesFactor object (possibly) containing the evidence from the data.

version

character string giving the version and revision number of the package that the model was created in

[Package BayesFactor version 0.9.12-4.2 Index]