BFprobability-class {BayesFactor}	R Documentation

General S4 class for representing multiple model probability comparisons

Description

The BFprobability class is a general S4 class for representing models model comparison via prior or posterior probabilities.

Usage


## S4 method for signature 'BFprobability,numeric'
e1 / e2

## S4 method for signature 'BFprobability,numeric'
e1 - e2

## S4 method for signature 'BFprobability,index,missing,missing'
x[i, j, ..., drop = TRUE]

## S4 method for signature 'BFprobability,character'
filterBF(x, name, perl = FALSE,
  fixed = FALSE, ...)

Arguments

e1
BFprobability object

e2
new normalization constant

x
BFprobability object

i
indices indicating elements to extract

j
unused for BFprobability objects

...
further arguments passed to related methods

drop
unused

name
regular expression to search name

perl
logical. Should perl-compatible regexps be used? See ?grepl for details.

fixed
logical. If TRUE, pattern is a string to be matched as is. See ?grepl for details.

Details

The BFprobability class has the following slots defined:

odds

A BFodds object containing the models from which to compute the probabilities

normalize

the sum of the probabilities of all models (will often be 1.0)

version

character string giving the version and revision number of the package that the model was created in

[Package BayesFactor version 0.9.12-4.2 Index]