BayesFactor-package {BayesFactor}	R Documentation

BayesFactor: Computation of Bayes Factors for Common Designs

Description

A suite of functions for computing various Bayes factors for simple designs, including contingency tables, one- and two-sample designs, one-way designs, general ANOVA designs, and linear regression.

This package contains function to compute Bayes factors for a number of research designs and hypotheses, including t tests, ANOVA, and linear regression, correlations, proportions, and contingency tables.

Details

Package: 	 BayesFactor

Type: 	 Package

Version: 	 0.9.12-4.8

Date: 	 2026-03-08

License: 	 GPL 2.0

LazyLoad: 	 yes

The following methods are currently implemented, with more to follow:

general linear models (including linear mixed effects models): generalTestBF, lmBF

linear regression: regressionBF, lmBF, linearReg.R2stat;

linear correlation: correlationBF;

t tests: ttestBF, ttest.tstat;

meta-analytic t tests: meta.ttestBF

ANOVA: anovaBF, lmBF, oneWayAOV.Fstat;

contingency tables: contingencyTableBF;

single proportions: proportionBF;

linear correlations: correlationBF;

Other useful functions: posterior, for sampling from posterior distributions; recompute, for re-estimating a Bayes factor or posterior distribution; compare, to compare two model posteriors; and plot.BFBayesFactor, for plotting Bayes factor objects.

Author(s)

Maintainer: Richard D. Morey richarddmorey@gmail.com [copyright holder]

Authors:

Jeffrey N. Rouder jrouder@uci.edu

Other contributors:

Tahira Jamil tahjamil@gmail.com [contributor, copyright holder]

Simon Urbanek simon.urbanek@r-project.org [contributor, copyright holder]

Karl Forner karl.forner@gmail.com [contributor, copyright holder]

Alexander Ly Alexander.Ly.NL@gmail.com [contributor, copyright holder]

Richard D. Morey and Jeffrey N. Rouder (with contributions from Tahira Jamil)

Maintainer: Richard D. Morey <richarddmorey@gmail.com>

References

Liang, F. and Paulo, R. and Molina, G. and Clyde, M. A. and Berger, J. O. (2008). Mixtures of g-priors for Bayesian Variable Selection. Journal of the American Statistical Association, 103, pp. 410-423

Rouder, J. N., Speckman, P. L., Sun, D., Morey, R. D., and Iverson, G. (2009). Bayesian t-tests for accepting and rejecting the null hypothesis. Psychonomic Bulletin & Review, 16, 225-237

Rouder, J. N., Morey, R. D., Speckman, P. L., Province, J. M., (2012) Default Bayes Factors for ANOVA Designs. Journal of Mathematical Psychology. 56. p. 356-374.

See Also

Useful links:

https://richarddmorey.github.io/BayesFactor/

Report bugs at https://github.com/richarddmorey/BayesFactor/issues

Examples

## See specific functions for examples.

[Package BayesFactor version 0.9.12-4.8 Index]