as.BFBayesFactor {BayesFactor}	R Documentation

Function to coerce objects to the BFBayesFactor class

Description

This function coerces objects to the BFBayesFactor class

Usage


as.BFBayesFactor(object)

Arguments

object
an object of appropriate class (for now, BFBayesFactorTop)

Details

Function to coerce objects to the BFBayesFactor class

Currently, this function will only work with objects of class BFBayesFactorTop, which are output from the functions anovaBF and regressionBF when the whichModels argument is set to 'top'

Value

An object of class BFBayesFactor

Author(s)

Richard D. Morey (richarddmorey@gmail.com)

See Also

regressionBF, anovaBF whose output is appropriate for use with this function when whichModels='top'

[Package BayesFactor version 0.9.12-4.2 Index]