as.BFprobability {BayesFactor}	R Documentation

Function to coerce objects to the BFprobability class

Description

This function coerces objects to the BFprobability class

Usage


as.BFprobability(object, normalize = NULL, lognormalize = NULL)

Arguments

object
an object of appropriate class (BFodds)

normalize
the sum of the probabilities for all models in the object (1 by default)

lognormalize
alternative to normalize; the logarithm of the normalization constant (0 by default)

Details

Function to coerce objects to the BFprobability class

Currently, this function will only work with objects of class BFOdds.

Value

An object of class BFprobability

Author(s)

Richard D. Morey (richarddmorey@gmail.com)

[Package BayesFactor version 0.9.12-4.2 Index]