enumerateGeneralModels {BayesFactor}	R Documentation

Function for generation of nested linear models

Description

Generate lists of nested models, given a model formula

Usage


enumerateGeneralModels(fmla, whichModels, neverExclude = NULL,
  includeBottom = TRUE, data = NULL)

Arguments

fmla
formula for the "full" model

whichModels
which subsets of models to generate

neverExclude
a character vector of terms to never remove

includeBottom
Include the base model containing only neverExclude terms

data
a data frame containing the columns mentioned in fmla

Details

This is a backend function not intended for users. It is exposed for third-party applications.

[Package BayesFactor version 0.9.12-4.2 Index]