filterBF {BayesFactor}	R Documentation

Filter the elements of an object according to some pre-specified criteria

Description

Filter the elements of an object according to some pre-specified criteria

Usage


filterBF(x, name, perl = FALSE, fixed = FALSE, ...)

Arguments

x
object

name
regular expression to search name

perl
logical. Should perl-compatible regexps be used? See ?grepl for details.

fixed
logical. If TRUE, pattern is a string to be matched as is. See ?grepl for details.

...
arguments passed to and from related methods

Value

Returns a filtered object

[Package BayesFactor version 0.9.12-4.2 Index]