%same% {BayesFactor}	R Documentation

Compare two objects to see if they are the 'same', for some loose definition of same

Description

Compare two objects to see if they are the 'same', for some loose definition of same

Usage


x %same% y

Arguments

x
first object

y
second object

Value

Returns TRUE or FALSE

[Package BayesFactor version 0.9.12-4.2 Index]