%termin% {BayesFactor}	R Documentation

Find a model term in a vector of model terms

Description

Find a model term in a vector of model terms

Usage


x %termin% table

Arguments

x
the terms to be matched

table
the terms to be matched against

Value

A logical vector of the same length as x, indicating if a match was located for each element of x.

[Package BayesFactor version 0.9.12-4.2 Index]