meta.ttestBF {BayesFactor}	R Documentation

Function for Bayesian analysis of one- and two-sample designs

Description

This function computes mata-analytic Bayes factors, or samples from the posterior, for one- and two-sample designs where multiple t values have been observed.

Usage


meta.ttestBF(t, n1, n2 = NULL, nullInterval = NULL, rscale = "medium",
  posterior = FALSE, callback = function(...) as.integer(0), ...)

Arguments

t
a vector of t statistics

n1
a vector of sample sizes for the first (or only) condition

n2
a vector of sample sizes. If NULL, a one-sample design is assumed

nullInterval
optional vector of length 2 containing lower and upper bounds of an interval hypothesis to test, in standardized units

rscale
prior scale. A number of preset values can be given as strings; see Details.

posterior
if TRUE, return samples from the posterior instead of Bayes factor

callback
callback function for third-party interfaces

...
further arguments to be passed to or from methods.

Details

The Bayes factor provided by meta.ttestBF tests the null hypothesis that the true effect size (or alternatively, the noncentrality parameters) underlying a set of t statistics is 0. Specifically, the Bayes factor compares two hypotheses: that the standardized effect size is 0, or that the standardized effect size is not 0. Note that there is assumed to be a single, common effect size delta underlying all t statistics. For one-sample tests, the standardized effect size is (mu-mu0)/sigma; for two sample tests, the standardized effect size is (mu2-mu1)/sigma.

A Cauchy prior is placed on the standardized effect size. The rscale argument controls the scale of the prior distribution, with rscale=1 yielding a standard Cauchy prior. See the help for ttestBF and the references below for more details.

The Bayes factor is computed via Gaussian quadrature. Posterior samples are drawn via independent-candidate Metropolis-Hastings.

Value

If posterior is FALSE, an object of class BFBayesFactor containing the computed model comparisons is returned. If nullInterval is defined, then two Bayes factors will be computed: The Bayes factor for the interval against the null hypothesis that the standardized effect is 0, and the corresponding Bayes factor for the compliment of the interval.

If posterior is TRUE, an object of class BFmcmc, containing MCMC samples from the posterior is returned.

Note

To obtain the same Bayes factors as Rouder and Morey (2011), change the prior scale to 1.

Author(s)

Richard D. Morey (richarddmorey@gmail.com)

References

Morey, R. D. & Rouder, J. N. (2011). Bayes Factor Approaches for Testing Interval Null Hypotheses. Psychological Methods, 16, 406-419

Rouder, J. N., Speckman, P. L., Sun, D., Morey, R. D., & Iverson, G. (2009). Bayesian t-tests for accepting and rejecting the null hypothesis. Psychonomic Bulletin & Review, 16, 225-237

Rouder, J. N. & Morey, R. D. (2011). A Bayes Factor Meta-Analysis of Bem's ESP Claim. Psychonomic Bulletin & Review, 18, 682-689

See Also

ttestBF

Examples


## Bem's (2010) data (see Rouder & Morey, 2011)
t=c(-.15,2.39,2.42,2.43)
N=c(100,150,97,99)

## Using rscale=1 and one-sided test to be
## consistent with Rouder & Morey (2011)
bf = meta.ttestBF(t, N, rscale=1, nullInterval=c(0, Inf))
bf[1]

## plot posterior distribution of delta, assuming alternative
## turn off progress bar for example
samples = posterior(bf[1], iterations = 1000, progress = FALSE)
## Note that posterior() respects the nullInterval
plot(samples)
summary(samples)

[Package BayesFactor version 0.9.12-4.2 Index]