BFmodel-class {BayesFactor}	R Documentation

General S4 classes for representing models for comparison

Description

The BFmodel is a general S4 class for representing models for comparison. The more classes BFlinearModel, BFindepSample, and BFoneSample inherit directly from BFmodel.

Details

These model classes all have the following slots defined:

type

Model type

identifier

a list uniquely identifying the model from other models of the same type

prior

list giving appropriate prior settings for the model

dataTypes

a character vector whose names are possible columns in the data; elements specify the corresponding data type, currently one of c("fixed","random","continuous")

shortName

a short, readable identifying string

longName

a longer, readable identifying string

analysis

object storing information about a previous analysis of this model

version

character string giving the version and revision number of the package that the model was created in

[Package BayesFactor version 0.9.12-4.2 Index]