newPriorOdds {BayesFactor}	R Documentation

Create prior odds from a Bayes factor object

Description

Create a prior odds object from a Bayes factor object

Usage


newPriorOdds(bf, type = "equal")

Arguments

bf
A BFBayesFactor object, eg, from an analysis

type
The type of prior odds to create (by default "equal"; see details)

Details

This function takes a Bayes factor object and, using its structure and specified type of prior odds, will create a prior odds object.

For now, the only type is "equal", which assigns equal prior odds to all models.

Value

A (prior) BFodds object, which can then be multiplied by the BFBayesFactor object to obtain posterior odds.

Author(s)

Richard D. Morey (richarddmorey@gmail.com)

[Package BayesFactor version 0.9.12-4.2 Index]