oneWayAOV.Fstat {BayesFactor}	R Documentation

Use F statistic to compute Bayes factor for balanced one-way designs

Description

Using the classical F test statistic for a balanced one-way design, this function computes the corresponding Bayes factor test.

Usage


oneWayAOV.Fstat(F, N, J, rscale = "medium", simple = FALSE)

Arguments

F
F statistic from classical ANOVA

N
number of observations per cell or group

J
number of cells or groups

rscale
numeric prior scale

simple
if TRUE, return only the Bayes factor

Details

For F statistics computed from balanced one-way designs, this function can be used to compute the Bayes factor testing the model that all group means are not equal to the grand mean, versus the null model that all group means are equal. It can be used when you don't have access to the full data set for analysis by lmBF, but you do have the test statistic.

For details about the model, see the help for anovaBF, and the references therein.

The Bayes factor is computed via Gaussian quadrature.

Value

If simple is TRUE, returns the Bayes factor (against the intercept-only null). If FALSE, the function returns a vector of length 3 containing the computed log(e) Bayes factor, along with a proportional error estimate on the Bayes factor and the method used to compute it.

Note

oneWayAOV.Fstat should only be used with F values obtained from balanced designs.

Author(s)

Richard D. Morey (richarddmorey@gmail.com)

References

Morey, R. D., Rouder, J. N., Pratte, M. S., \& Speckman, P. L. (2011). Using MCMC chain outputs to efficiently estimate Bayes factors. Journal of Mathematical Psychology, 55, 368-378

See Also

integrate, aov; see lmBF for the intended interface to this function, using the full data set.

Examples


## Example data "InsectSprays" - see ?InsectSprays
require(stats); require(graphics)
boxplot(count ~ spray, data = InsectSprays, xlab = "Type of spray", 
        ylab = "Insect count", main = "InsectSprays data", varwidth = TRUE, 
        col = "lightgray")

## Classical analysis (with transformation)
classical <- aov(sqrt(count) ~ spray, data = InsectSprays)
plot(classical)
summary(classical)

## Bayes factor (a very large number)
Fvalue <- anova(classical)$"F value"[1]
result <- oneWayAOV.Fstat(Fvalue, N=12, J=6)
exp(result[['bf']])

[Package BayesFactor version 0.9.12-4.2 Index]