options-BayesFactor {BayesFactor}	R Documentation

options() for package BayesFactor

Description

Options that can be set for the BayesFactor package

Details

The BayesFactor package has numerous options that can be set to globally change the behavior of the functions in the package. These options can be changed using options().

BFMaxModels

Integer; maximum number of models to analyze in anovaBF or regressionBF

BFprogress

If TRUE, progress bars are on by default; if FALSE, they are disabled by default.

BFpretestIterations

Integer; if sampling is needed to compute the Bayes factor, the package attempts to choose the most efficient sampler. This option controls the number of initial test iterations.

BFapproxOptimizer

"nlm" or "optim"; changes the optimization function used for the importance sampler. If one fails, try the other.

BFapproxLimits

Vector of length two containing the lower and upper limits on on log(g) before the the posterior returns -Inf. This only affects the initial optimization step for the importance sampler.

BFfactorsMax

Maximum number of factors to try to do enumeration with in generalTestBF.

BFcheckProbabilityList

Check for duplicate models when creating BFprobability objects?

See Also

options

[Package BayesFactor version 0.9.12-4.2 Index]