plot.BFBayesFactor {BayesFactor}	R Documentation

Plot a Bayes factor object

Description

Plot a Bayes factor object

Usage


## S3 method for class 'BFBayesFactor'
plot(x, include1 = TRUE, addDenom = FALSE,
  sortbf = TRUE, logbase = c("log10", "log2", "ln"), marginExpand = 0.4,
  cols = c("wheat", "lightslateblue"), main = paste("vs.",
  x@denominator@longName), pars = NULL, ...)

Arguments

x
a BFBayesFactor object

include1
if TRUE, ensure that Bayes factor = 1 is on the plot

addDenom
if TRUE, add the denominator model into the group

sortbf
sort the Bayes factors before plotting them? Defaults to TRUE

logbase
the base of the log Bayes factors in the plot

marginExpand
an expansion factor for the left margin, in case more space is needed for model names

cols
a vector of length two of valid color names or numbers

main
a character vector for the plot title

pars
a list of par() settings

...
additional arguments to pass to barplot()

Details

This function creates a barplot of the (log) Bayes factors in a Bayes factor object. Error bars are added (though in many cases they may be too small to see) in red to show the error in estimation of the Bayes factor. If a red question mark appears next to a bar, then that Bayes factor has no error estimate available.

Author(s)

Richard D. Morey (richarddmorey@gmail.com)

Examples


data(puzzles)

bfs = anovaBF(RT ~ shape*color + ID, data = puzzles, whichRandom="ID", progress=FALSE)
plot(bfs)

[Package BayesFactor version 0.9.12-4.2 Index]