ttest.tstat {BayesFactor}	R Documentation

Use t statistic to compute Bayes factor for one- and two- sample designs

Description

Using the classical t test statistic for a one- or two-sample design, this function computes the corresponding Bayes factor test.

Usage


ttest.tstat(t, n1, n2 = 0, nullInterval = NULL, rscale = "medium",
  complement = FALSE, simple = FALSE)

Arguments

t
classical t statistic

n1
size of first group (or only group, for one-sample tests)

n2
size of second group, for independent-groups tests

nullInterval
optional vector of length 2 containing lower and upper bounds of an interval hypothesis to test, in standardized units

rscale
numeric prior scale

complement
if TRUE, compute the Bayes factor against the complement of the interval

simple
if TRUE, return only the Bayes factor

Details

This function can be used to compute the Bayes factor corresponding to a one-sample, a paired-sample, or an independent-groups t test, using the classical t statistic. It can be used when you don't have access to the full data set for analysis by ttestBF, but you do have the test statistic.

For details about the model, see the help for ttestBF, and the references therein.

The Bayes factor is computed via Gaussian quadrature.

Value

If simple is TRUE, returns the Bayes factor (against the null). If FALSE, the function returns a vector of length 3 containing the computed log(e) Bayes factor, along with a proportional error estimate on the Bayes factor and the method used to compute it.

Note

In version 0.9.9, the behaviour of this function has changed in order to produce more uniform results. In version 0.9.8 and before, this function returned two Bayes factors when nullInterval was non-NULL: the Bayes factor for the interval versus the null, and the Bayes factor for the complement of the interval versus the null. Starting in version 0.9.9, in order to get the Bayes factor for the complement, it is required to set the complement argument to TRUE, and the function only returns one Bayes factor.

Author(s)

Richard D. Morey (richarddmorey@gmail.com) and Jeffrey N. Rouder (rouderj@missouri.edu)

References

Morey, R. D. & Rouder, J. N. (2011). Bayes Factor Approaches for Testing Interval Null Hypotheses. Psychological Methods, 16, 406-419

Rouder, J. N., Speckman, P. L., Sun, D., Morey, R. D., & Iverson, G. (2009). Bayesian t-tests for accepting and rejecting the null hypothesis. Psychonomic Bulletin & Review, 16, 225-237

See Also

integrate, t.test; see ttestBF for the intended interface to this function, using the full data set.

Examples


## Classical example: Student's sleep data
data(sleep)
plot(extra ~ group, data = sleep)

## t.test() gives a t value of -4.0621
t.test(extra ~ group, data = sleep, paired=TRUE)
## Gives a Bayes factor of about 15
## in favor of the alternative hypothesis
result <- ttest.tstat(t = -4.0621, n1 = 10)
exp(result[['bf']])

[Package BayesFactor version 0.9.12-4.2 Index]